rpi-pwm: Make 'audio' mode more robust
Because messing around with various PWM registers can be dangerous when
playing audio, this patch will switch the driver into "delayed" updating
when placed into audio mode.  This means that writes to "degrees",
"percent", and "frequency" will not take effect immediately.

It also allows you to enable or disable audio-out on the PWM pin by
toggling the value of "active": to reflect audio on the PWM pin, switch to
"audio" mode and write a 1 into "active".  To silence audio on the PWM
pin, write a 0 into "active".