rpi-pwm: Rename 'degrees' to 'servo', use scale
Different PWM motors have a different number of steps available.  Rather
than going from 0-100 (percentages) or 0-360 ("degrees"), we allow the
user to select the number of steps their motor supports, and then use that
scale.

To set the maximum number of steps, write a value out to servo_max.  The
value in "servo" will get scaled accordingly (and with very large or very
small nubmers of steps, "servo" might get rounded badly.)

Note that because of the way the math works out, very large numbers of
steps (e.g. in the 100k+ range) will not work, as the values will
overflow.