fbtft: Add a sysfs parameter to set the vertical scrolling position on a window.
Add a sysfs param scroll_pos which sets the scanline where the display
starts for 'rolling scrolling'. This needs support from the specific
display driver as well. When supported it can be used to implement smooth
vertical (when the display is in portrait mode) hor horizontal (when the
display is used in landscape mode) scrolling by strategically changing the
start scanline of the display and erasing/redrawing only the wrapping part.
This way you never need to redraw what's already there, preventing slow
full screen redraws over spi.